Add alpha optimisation to the minimax
This is a little difficult to explain, you may want to look on wikipedia
to have a better explanation than mine... I'll try my best.
Alpha optimisation is a way to avoid useless caluclation. When the
'player' is searching for his next move we pass it an alpha which
correspond to the minimal interresting value, if the player,
searching for the maximal move find a score superior to alpha, then he
should stop and return anything with a score >alpha.